+++ Inventario/Application/ViewModels/FeeRuleViewModel.cs
@@ -12,6 +12,6 @@ namespace Application.ViewModels
         public ProductViewModel Product { get; set; }
         public DateTime Date { get; set; }
         public int FeesAmountTo { get; set; }
-        public double Percentage { get; set; }
+        public decimal Percentage { get; set; }
     }
 }
