+++ Inventario/Application/ViewModels/Filters/FeeRuleFiltersViewModel.cs
@@ -9,7 +9,7 @@ namespace Application.ViewModels
     {
         public Guid? ProductId { get; set; }
         public int? FeesAmountTo { get; set; }
-        public double? Percentage { get; set; }
+        public decimal? Percentage { get; set; }
 
         public bool IsEmpty()
         {
