+++ Inventario/Application/ViewModels/ProductForCreationViewModel.cs
@@ -14,8 +14,8 @@ namespace Application.ViewModels
         public Guid CategoryId { get; set; }
         public Guid? VendorId { get; set; }
         public string Brand { get; set; }
-        public double? MinimumStock { get; set; }
-        public double Stock { get; set; }
+        public decimal? MinimumStock { get; set; }
+        public decimal Stock { get; set; }
         public string UnitOfMeasurement { get; set; }
         public PriceForCreationViewModel PurchasePrice { get; set; }
         public PriceForCreationViewModel SalePrice { get; set; }
