+++ Inventario/Logic/Dtos/FeeRuleForCreationDto.cs
@@ -8,6 +8,6 @@ namespace Logic.Dtos
     {
         public Guid ProductId { get; set; }
         public int FeesAmountTo { get; set; }
-        public double Percentage { get; set; }
+        public decimal Percentage { get; set; }
     }
 }
