+++ Inventario/Logic/Dtos/Payments/CashForCreationDto.cs
@@ -7,6 +7,6 @@ namespace Logic.Dtos
 {
     public class CashForCreationDto : PaymentForCreationDto
     {
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
     }
 }
