+++ Inventario/Logic/Dtos/ProductForCreationDto.cs
@@ -13,8 +13,8 @@ namespace Logic.Dtos
         public Guid CategoryId { get; set; }
         public Guid VendorId { get; set; }
         public string Brand { get; set; }
-        public double? MinimumStock { get; set; }
-        public double Stock { get; set; }
+        public decimal? MinimumStock { get; set; }
+        public decimal Stock { get; set; }
         public string UnitOfMeasurement { get; set; }
         public PriceForCreationDto PurchasePrice { get; set; }
         public PriceForCreationDto SalePrice { get; set; }
